Redux Ducks Refactor * Refactor Config and Loader tabs Redux-Sagas flow to use the [Ducks pattern](https://github.com/erikras/ducks-modular-redux) * This moves code scattered between `actions/`, `reducers/` and `sagas/` directories into a single combined file in a new `ducks/` directory * This should make it conceptually easier to understand the flow and have all related functionality together in a single file for easier refactoring (not switching between lots of files) * [CLOSES #401]
