--- misk/web/tabs/config/src/actions/types.ts
@@ -1,9 +1,0 @@
-export interface IActionType {
-  CONFIG: CONFIG
-}
-
-export enum CONFIG {
-  FAILURE = "CONFIG_FAILURE",
-  GET_ALL = "CONFIG_GET_ALL",
-  SUCCESS = "CONFIG_SUCCESS"
-}
