+++ misk/web/tabs/config/src/containers/TabContainer.tsx
@@ -1,9 +1,8 @@
 import { OfflineComponent } from "@misk/components"
 import * as React from "react"
 import { connect } from "react-redux"
-import { dispatchConfig } from "../actions"
 import { ConfigComponent } from "../components"
-import { IState } from "../reducers"
+import { dispatchConfig, IState } from "../ducks"
 
 interface ITabProps {
   config: {
@@ -18,11 +17,11 @@ export interface IConfigResource {
   file: string
 }
 
-const configUrl = "/api/config/all"
+const apiUrl = "/api/config/all"
 
 class TabContainer extends React.Component<ITabProps, {children : any}> {
   componentDidMount() {
-    this.props.getAllConfig(configUrl)
+    this.props.getAllConfig(apiUrl)
   }
 
   render() {
@@ -33,7 +32,7 @@ class TabContainer extends React.Component<ITabProps, {children : any}> {
       )
     } else {
       return (
-        <OfflineComponent title={"Error Loading Config Data"} endpoint={configUrl}/>
+        <OfflineComponent title={"Error Loading Config Data"} endpoint={apiUrl}/>
       )
     }
   }
