+++ misk/web/tabs/config/src/index.tsx
@@ -9,8 +9,7 @@ import { Provider } from "react-redux"
 import { applyMiddleware, compose, createStore } from "redux"
 import createSagaMiddleware from "redux-saga"
 import App from "./App"
-import rootReducer from "./reducers"
-import rootSaga from "./sagas"
+import { rootReducer, rootSaga } from "./ducks"
 
 export * from "./components"
 export * from "./containers"
@@ -58,7 +57,7 @@ if (module.hot) {
   })
 
   // Reload reducers
-  module.hot.accept("./reducers", () => {
+  module.hot.accept("./ducks", () => {
     store.replaceReducer(connectRouter(history)(rootReducer))
   })
 }
