--- misk/web/tabs/config/src/sagas/index.ts
@@ -1,8 +1,0 @@
-import { all, fork } from "redux-saga/effects"
-import watchConfigSagas from "./configSaga"
-
-export default function * rootSaga () {
-  yield all([
-    fork(watchConfigSagas)
-  ])
-}
