+++ misk/web/tabs/loader/src/App.tsx
@@ -1,5 +1,5 @@
 import { ConnectedRouter } from "connected-react-router"
-import {History} from "history"
+import { History } from "history"
 import * as React from "react"
 import routes from "./routes"
