+++ misk/web/tabs/loader/src/ducks/item.ts
@@ -1,17 +1,70 @@
-// In case you need to use a selector
-// import also select from redux-saga/effects
-// and then simplie yield select(yourSelector())
-//
-// In case you need to redirect to whatever route
-// import { push } from react-router-redux and then
-// yield put(push('/next-page'))
-
+import { createAction, IMiskAction } from "@misk/common"
 import axios from "axios"
+import { fromJS, List } from "immutable"
 import { all, call, put, takeLatest } from "redux-saga/effects"
 
-import {
-  dispatchItem, IMiskAction, IActionType, ITEM
-} from "../actions"
+/**
+ * Actions
+ */
+interface IActionType {
+  ITEM: ITEM
+}
+
+export enum ITEM {
+  GET = "ITEM_GET",
+  GET_ONE = "ITEM_GET_ONE",
+  SAVE = "ITEM_SAVE",
+  PUT = "ITEM_PUT",
+  PATCH = "ITEM_PATCH",
+  DELETE = "ITEM_DELETE",
+  SUCCESS = "ITEM_SUCCESS",
+  FAILURE = "ITEM_FAILURE"
+}
+
+export const dispatchItem = {
+  delete: (id: number) => createAction(ITEM.DELETE, { id, loading: true, success: false, error: null }),
+  failure: (error: any) => createAction(ITEM.FAILURE, { ...error, loading: false, success: false }),
+  patch: (id: number, data: any) => createAction(ITEM.PATCH, { id, ...data, loading: true, success: false, error: null }),
+  put: (id: number, data: any) => createAction(ITEM.PUT, { id, ...data, loading: true, success: false, error: null }),
+  request: () => createAction(ITEM.GET, { loading: true, success: false, error: null }),
+  requestOne: (id: number) => createAction(ITEM.GET_ONE, { id, loading: true, success: false, error: null }),
+  save: (data: any) => createAction(ITEM.SAVE, { ...data, loading: true, success: false, error: null }),
+  success: (data: any) => createAction(ITEM.SUCCESS, { ...data, loading: false, success: true, error: null }),
+}
+
+/**
+ * Reducer
+ * @param state 
+ * @param action 
+ */
+export interface IItemState {
+  [key: string]: any
+}
+
+const initialState = fromJS({
+  data: List([]),
+  error: null,
+  loading: false,
+  success: false,
+})
+
+export default function ItemReducer (state = initialState, action: IMiskAction<string, {}>) {
+  switch (action.type) {
+    case ITEM.GET:
+    case ITEM.GET_ONE:
+    case ITEM.SAVE:
+    case ITEM.DELETE:
+    case ITEM.SUCCESS:
+    case ITEM.FAILURE:
+      return state.merge(action.payload)
+    default:
+      return state
+  }
+}
+
+/**
+ * Sagas
+ */
 
 function * handleGet () {
   try {
@@ -72,7 +125,7 @@ function * handleDelete (action: IMiskAction<IActionType, {id: number}>) {
   }
 }
 
-function * watchItemSagas () {
+export function * watchItemSagas () {
   yield all([
     takeLatest(ITEM.GET, handleGet),
     takeLatest(ITEM.GET_ONE, handleGetOne),
@@ -82,5 +135,3 @@ function * watchItemSagas () {
     takeLatest(ITEM.DELETE, handleDelete)
   ])
 }
-
-export default watchItemSagas
